OvmfPkg ResetSystemLib: Fix build issue with VS2005
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@11276 6f19259b-4bc3-4df7-8a09-765794883524

+++ OvmfPkg/Library/ResetSystemLib/ResetSystemLib.c
@@ -25,7 +25,7 @@ AcpiPmControl (
 {
   ASSERT (SuspendType < 6);
 
-  IoAndThenOr16 (0x404, ~0x3c00, SuspendType << 10);
+  IoAndThenOr16 (0x404, (UINT16) ~0x3c00, (UINT16) (SuspendType << 10));
   IoOr16 (0x404, BIT13);
   CpuDeadLoop ();
 }
